feat: 型定義をlib/types.tsに分離
- lib/types.ts を新設し型定義をmock-data.tsから分離
- mock-data.tsはデータのみに（lib/typesからimport）
- 6スクリーンの型importをlib/mock-dataからlib/typesに変更

Closes #59
